Apply Debian patches to binutils

+++ src-projects/binutils-2.16.1/gas/config/tc-h8300.c
@@ -529,7 +529,7 @@ static int
 constant_fits_width_p (struct h8_op *operand, unsigned int width)
 {
   return ((operand->exp.X_add_number & ~width) == 0
-	  || (operand->exp.X_add_number | width) == (unsigned)(~0));
+	  || (operand->exp.X_add_number | width) == (unsigned long)(~0));
 }
 
 static int
